Better Images + Buttons (read full)
Prep for the side panel update

When you add new images, please try to add them to the Images class. This will keep all the images in one place and only loaded once. Also faster to get it multiple times.

Old images still work, but please try to convert them when you see them and have time.

Same with the button class. They are more efficient, easier to change, and easier to make.
I changed the increment and decrement buttons because their definition was obvious, and will slowly work on converting other buttons.

Note to self and others: be sure to set buttons' active boolean to true or false when you need them to be active. There is likely a better way to do this, but I don't remember at this moment.

Added some //PLZ comments that where i need clarifacation, respond below the comments so i can see them when I get to editing that specific part of code. maybe with a //ANS

Thank you for reading, theres a reason this big comments section exists.